requirements.txt
absl-py==0.11.0
astor==0.8.1
bert-serving-client==1.10.0
bert-serving-server==1.10.0
cached-property==1.5.2
click==7.1.2
gast==0.4.0
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.35.0
h5py==3.1.0
importlib-metadata==3.4.0
joblib==1.0.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
nltk==3.5
numpy==1.20.0
pandas==1.2.1
protobuf==3.14.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.2
regex==2020.11.13
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.56.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
